+++ Sources/Scru128/Scru128Generator.swift
@@ -8,19 +8,18 @@ let defaultRollbackAllowance: UInt64 = 10_000  // 10 seconds
 ///
 /// The generator offers four different methods to generate a SCRU128 ID:
 ///
-/// | Flavor                                                 | Timestamp | Thread- | On big clock rewind |
-/// | ------------------------------------------------------ | --------- | ------- | ------------------- |
-/// | ``generate()``                                         | Now       | Safe    | Rewinds state       |
-/// | ``generateNoRewind()``                                 | Now       | Safe    | Returns `nil`       |
-/// | ``generateCore(_:)``                                   | Argument  | Unsafe  | Rewinds state       |
-/// | ``generateCoreNoRewind(timestamp:rollbackAllowance:)`` | Argument  | Unsafe  | Returns `nil`       |
+/// | Flavor                                                | Timestamp | Thread- | On big clock rewind |
+/// | ----------------------------------------------------- | --------- | ------- | ------------------- |
+/// | ``generate()``                                        | Now       | Safe    | Resets generator    |
+/// | ``generateOrAbort()``                                 | Now       | Safe    | Returns `nil`       |
+/// | ``generateOrResetCore(timestamp:rollbackAllowance:)`` | Argument  | Unsafe  | Resets generator    |
+/// | ``generateOrAbortCore(timestamp:rollbackAllowance:)`` | Argument  | Unsafe  | Returns `nil`       |
 ///
-/// Each method returns monotonically increasing IDs unless a `timestamp` provided is significantly
-/// (by ten seconds or more by default) smaller than the one embedded in the immediately preceding
-/// ID. If such a significant clock rollback is detected, the `generate` method rewinds the
-/// generator state and returns a new ID based on the current `timestamp`, whereas the experimental
-/// `NoRewind` variants keep the state untouched and return `nil`. `Core` functions offer low-level
-/// thread-unsafe primitives.
+/// All of these methods return monotonically increasing IDs unless a `timestamp` provided is
+/// significantly (by default, ten seconds or more) smaller than the one embedded in the immediately
+/// preceding ID. If such a significant clock rollback is detected, the `generate` (OrReset) method
+/// resets the generator and returns a new ID based on the given `timestamp`, while the `OrAbort`
+/// variants abort and return `nil`. The `Core` functions offer low-level thread-unsafe primitives.
 public class Scru128Generator {
   private var timestamp: UInt64 = 0
   private var counterHi: UInt32 = 0
@@ -33,20 +32,11 @@ public class Scru128Generator {
   ///
   /// Note that the generator object should be protected from concurrent accesses during the
   /// sequential calls to a generation method and this property to avoid race conditions.
-  ///
-  /// Examples:
-  ///
-  /// ```swift
-  /// let g = Scru128Generator()
-  /// let x = g.generate()
-  /// let y = g.generate()
-  /// precondition(
-  ///   g.lastStatus != Scru128Generator.Status.clockRollback,
-  ///   "clock moved backward"
-  /// )
-  /// assert(x < y)
-  /// ```
-  private(set) public var lastStatus: Status = Status.notExecuted
+  @available(*, deprecated, message: "Use `generateOrAbort()` to guarantee monotonicity.")
+  public var lastStatus: Status { lastStatusInternal }
+
+  /// For internal use to supress deprecation warnings
+  internal var lastStatusInternal: Status = Status.notExecuted
 
   /// The random number generator used by the generator.
   private var rng: RandomNumberGenerator
@@ -64,90 +54,116 @@ public class Scru128Generator {
     self.rng = rng
   }
 
-  /// Generates a new SCRU128 ID object from the current `timestamp`.
+  /// Generates a new SCRU128 ID object from the current `timestamp`, or resets the generator upon
+  /// significant timestamp rollback.
   ///
   /// See the ``Scru128Generator`` class documentation for the description.
   public func generate() -> Scru128Id {
     lock.lock()
     defer { lock.unlock() }
-    return generateCore(UInt64(Date().timeIntervalSince1970 * 1_000))
+    return generateOrResetCore(
+      timestamp: UInt64(Date().timeIntervalSince1970 * 1_000),
+      rollbackAllowance: defaultRollbackAllowance)
+
   }
 
-  /// _Experimental_. Generates a new SCRU128 ID object from the current `timestamp`, guaranteeing
-  /// the monotonic order of generated IDs despite a significant timestamp rollback.
+  /// Generates a new SCRU128 ID object from the current `timestamp`, or returns `nil` upon
+  /// significant timestamp rollback.
   ///
   /// See the ``Scru128Generator`` class documentation for the description.
-  public func generateNoRewind() -> Scru128Id? {
+  ///
+  /// Examples:
+  ///
+  /// ```swift
+  /// let g = Scru128Generator()
+  /// let x = g.generateOrAbort()!
+  /// guard let y = g.generateOrAbort() else {
+  ///   fatalError("The clock went backwards by ten seconds!")
+  /// }
+  /// assert(x < y)
+  /// ```
+  public func generateOrAbort() -> Scru128Id? {
     lock.lock()
     defer { lock.unlock() }
-    return generateCoreNoRewind(
+    return generateOrAbortCore(
       timestamp: UInt64(Date().timeIntervalSince1970 * 1_000),
       rollbackAllowance: defaultRollbackAllowance)
   }
 
-  /// Generates a new SCRU128 ID object from the `timestamp` passed.
-  ///
-  /// See the ``Scru128Generator`` class documentation for the description.
-  ///
-  /// Unlike ``generate()``, this method is NOT thread-safe. The generator object should be
-  /// protected from concurrent accesses using a mutex or other synchronization mechanism to avoid
-  /// race conditions.
-  ///
-  /// - Precondition: `timestamp` must be a 48-bit positive integer.
-  public func generateCore(_ timestamp: UInt64) -> Scru128Id {
-    let rollbackAllowance = defaultRollbackAllowance
-    if let value = generateCoreNoRewind(timestamp: timestamp, rollbackAllowance: rollbackAllowance)
-    {
-      return value
-    } else {
-      // reset state and resume
-      self.timestamp = 0
-      tsCounterHi = 0
-      let value = generateCoreNoRewind(timestamp: timestamp, rollbackAllowance: rollbackAllowance)!
-      lastStatus = Status.clockRollback
-      return value
-    }
-  }
-
-  /// _Experimental_. Generates a new SCRU128 ID object from the `timestamp` passed, guaranteeing
-  /// the monotonic order of generated IDs despite a significant timestamp rollback.
+  /// Generates a new SCRU128 ID object from the `timestamp` passed, or resets the generator upon
+  /// significant timestamp rollback.
   ///
   /// See the ``Scru128Generator`` class documentation for the description.
   ///
   /// The `rollbackAllowance` parameter specifies the amount of `timestamp` rollback that is
   /// considered significant. A suggested value is `10_000` (milliseconds).
   ///
-  /// Unlike ``generateNoRewind()``, this method is NOT thread-safe. The generator object should be
+  /// Unlike ``generate()``, this method is NOT thread-safe. The generator object should be
   /// protected from concurrent accesses using a mutex or other synchronization mechanism to avoid
   /// race conditions.
   ///
   /// - Precondition: `timestamp` must be a 48-bit positive integer.
-  public func generateCoreNoRewind(timestamp: UInt64, rollbackAllowance: UInt64) -> Scru128Id? {
+  public func generateOrResetCore(timestamp: UInt64, rollbackAllowance: UInt64) -> Scru128Id {
+    if let value = generateOrAbortCore(timestamp: timestamp, rollbackAllowance: rollbackAllowance) {
+      return value
+    } else {
+      // reset state and resume
+      self.timestamp = 0
+      tsCounterHi = 0
+      let value = generateOrAbortCore(timestamp: timestamp, rollbackAllowance: rollbackAllowance)!
+      lastStatusInternal = Status.clockRollback
+      return value
+    }
+  }
+
+  /// A deprecated synonym for `generateOrResetCore(timestamp: timestamp, rollbackAllowance: 10_000)`.
+  @available(
+    *, deprecated,
+    message: "Use `generateOrResetCore(timestamp: timestamp, rollbackAllowance: 10_000)` instead."
+  )
+  public func generateCore(_ timestamp: UInt64) -> Scru128Id {
+    return generateOrResetCore(timestamp: timestamp, rollbackAllowance: defaultRollbackAllowance)
+  }
+
+  /// Generates a new SCRU128 ID object from the `timestamp` passed, or returns `nil` upon
+  /// significant timestamp rollback.
+  ///
+  /// See the ``Scru128Generator`` class documentation for the description.
+  ///
+  /// The `rollbackAllowance` parameter specifies the amount of `timestamp` rollback that is
+  /// considered significant. A suggested value is `10_000` (milliseconds).
+  ///
+  /// Unlike ``generateOrAbort()``, this method is NOT thread-safe. The generator object should be
+  /// protected from concurrent accesses using a mutex or other synchronization mechanism to avoid
+  /// race conditions.
+  ///
+  /// - Precondition: `timestamp` must be a 48-bit positive integer.
+  public func generateOrAbortCore(timestamp: UInt64, rollbackAllowance: UInt64) -> Scru128Id? {
     precondition(timestamp != 0 && timestamp <= maxTimestamp)
     precondition(rollbackAllowance <= maxTimestamp)
 
     if timestamp > self.timestamp {
       self.timestamp = timestamp
       counterLo = rng.next() & maxCounterLo
-      lastStatus = Status.newTimestamp
+      lastStatusInternal = Status.newTimestamp
     } else if timestamp + rollbackAllowance > self.timestamp {
       // go on with previous timestamp if new one is not much smaller
       counterLo += 1
-      lastStatus = Status.counterLoInc
+      lastStatusInternal = Status.counterLoInc
       if counterLo > maxCounterLo {
         counterLo = 0
         counterHi += 1
-        lastStatus = Status.counterHiInc
+        lastStatusInternal = Status.counterHiInc
         if counterHi > maxCounterHi {
           counterHi = 0
           // increment timestamp at counter overflow
           self.timestamp += 1
           counterLo = rng.next() & maxCounterLo
-          lastStatus = Status.timestampInc
+          lastStatusInternal = Status.timestampInc
         }
       }
     } else {
-      // abort if clock moves back to unbearable extent
+      // abort if clock went backwards to unbearable extent
       return nil
     }
 
@@ -159,7 +175,7 @@ public class Scru128Generator {
     return Scru128Id(self.timestamp, counterHi, counterLo, rng.next())
   }
 
-  /// The status code returned by ``lastStatus`` property.
+  /// _Deprecated_: The status code returned by ``lastStatus`` property.
   public enum Status {
     /// Indicates that the generator has yet to generate an ID.
     case notExecuted