+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2.4.0 - 2023-03-22
+
+### Added
+
+- `generateOrAbort()` and `generateOrAbortCore()` to `Scru128Generator`
+  (formerly named as `generateNoRewind()` and `generateCoreNoRewind()`)
+- `Scru128Generator#generateOrResetCore()`
+
+### Deprecated
+
+- `Scru128Generator#generateCore()`
+- `Scru128Generator#lastStatus` and `Scru128Generator.Status`
+
 ## v2.3.1 - 2023-03-19
 
 ### Added